docs: add CircleCI repo subscription note to publishing docs (#312) (#334)
you need to be a meber of an authorized github team and following the builds on circleci to be able to publish changes. Otherwise the CI build will fail with an "unauthorized" warning. It is unituitive that you have to specifically go and follow the circleci build but without that it seems that circle does not recognise the user trigger the build as authorized.