chore: fix ci. pin deps
a dep started breaking the build with errors of the form

```
npm ERR! Invalid dependency type requested: alias
```

which it turns out is a feature only supported in npm >= 6.9.

So this PR updeates the docker image and pins the deps by including the package-lock.json.

If you have a local package-lock.json, delete it, and accept this one into your heart.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>